+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+### 0.9.3 - 2016-08-24
+> renamed jobSchedule method to scheduleTask for proper functionality and readability
+
+> changed endpoint for scheduleTask to POST /:customer/task/schedule
+
+> added method getTaskSchedule(task_id, cb) pointing to GET /:customer/task/:task/schedule
+
 ### 0.9.2 - 2016-08-22
 
 > remove unused methods. (task and resource)